+++ package.json
@@ -13,6 +13,8 @@
     "@tailwindcss/vite": "^4.1.11",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router": "^7.7.0",
+    "react-router-dom": "^7.7.0",
     "tailwindcss": "^4.1.11"
   },
   "devDependencies": {
