+++ src/ProductList.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import ProductCard from "./ProductCard";
 
-function ProductList({products}){
+function ProductList({products=[]}){
     return (
         <div className="bg-white m-12 mt-0 mb-0  p-4 flex flex-wrap justify-center gap-2">
             {products.map(function(item){
                 return (
                 <ProductCard
                     key={item.title}
+                    id={item.id}
                     image={item.image}
                     title={item.title}
                     category={item.category}
