+++ src/alldata.js
@@ -1,0 +1,66 @@
+const alldata=[
+    {
+      id:1,
+      image:"https://www.arcprint.in/assets/media/products_common_imgs/black-coffee-mugs/description.jpg",
+      title:"Black Printed Coffee Mug",
+      category:"Mugs",
+      price:"15.00",
+    },
+    {
+      id:2,
+      image:"https://m.media-amazon.com/images/I/61rc196ZwUL._UF894,1000_QL80_.jpg",
+      title:"Father's day coffee mug",
+      category:"Mugs",
+      price:"19.00",
+    },
+    {
+      id:3,
+      image:"https://putra.in/cdn/shop/files/IMG-0022_1200x.jpg?v:1732647723",
+      title:"Green Printed Tshirt",
+      category:"Tshirts",
+      price:"34.00",
+    },
+    {
+      id:4,
+      image:"https://betweenboxes.in/cdn/shop/files/cust-mug-600x600-500x500.webp?v:1720606037",
+      title:"Personalised mug",
+      category:"Mugs",
+      price:"15.00",
+    },
+    {
+      id:5,
+      image:"https://www.botnia.in/cdn/shop/files/47_b87415d3-951b-4172-aa68-89d63ec8f5f0_800x.png?v:1696434833",
+      title:"Printed Brown Tshirts",
+      category:"Tshirts",
+      price:"25.00",
+    },
+    {
+      id:6,
+      image:"https://www.rockit.co.in/cdn/shop/files/2240101920-1-38_1.jpg?v:1715680470",
+      title:"Printed Dark Blue Tshirts",
+      category:"Tshirts",
+      price:"34.00",
+    },
+    {
+      id:7,
+      image:"https://muselot.in/cdn/shop/products/Never-give-up-1.jpg?v:1644528299",
+      title:"Printed Green tshirts",
+      category:"Tshirts",
+      price:"28.00",
+    },
+    {
+      id:8,
+      image:"https://muselot.in/cdn/shop/products/I-am-coffee-hot4.jpg?v:1625002565",
+      title:"Printed Tshirt Coffee Black Color",
+      category:"Tshirts",
+      price:"25.00",
+    },
+    {
+      id:9,
+      image:"https://images.meesho.com/images/products/248447273/hlsee_512.webp",
+      title:"Typography Teal Printed Tshirt",
+      category:"Tshirts",
+      price:"32.00",
+    },
+]
+export default alldata;
